next level.

+++ bo/qt-application/larch/bo/qt/start.py
@@ -2,6 +2,7 @@
 import os
 import sys
 import argparse
+import socket
 from pathlib import Path
 
 
@@ -18,6 +19,15 @@ def local_parser():
     return parser
 
 
+def get_free_port():
+    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    s.bind(("", 0))
+    s.listen(1)
+    port = s.getsockname()[1]
+    s.close()
+    return port
+
+
 def run(root, config=None, application=None):
     arg_parser = local_parser()
     args = arg_parser.parse_known_args()[0]
@@ -26,7 +36,7 @@ def run(root, config=None, application=None):
     except AttributeError:
         type_ = "local"   # a frozen version without this option
 
-    if type_ in ("compile", "server", "debug"):
+    if type_ in ("compile", "server"):
         from larch.bo.server import run
         config = config or {}
         config["wsheartbeat"] = 0  # websocket heartbeat not needed
@@ -35,25 +45,33 @@ def run(root, config=None, application=None):
 
     if os.environ.get("LARCH_BO_QT_FRONTEND_STARTED"):
         from larch.bo.server import run
-        from .pointcuts import InformFrontend
-        config.setdefault("application_pointcuts", []).append(InformFrontend)
-        sys.argv.append("--port=0")
         return run(root, config, application)
     else:
-        from subprocess import Popen, PIPE
+        from subprocess import Popen
         from .window import start_frontend
         config = config or {}
         config["args"] = args
 
-        config.setdefault("resources_path", Path(
+        resource_path = config.setdefault("resource_path", Path(
             config.get("build_path") or Path.cwd().resolve()/".lfrontend")/"dist")
 
-        args = [sys.executable] + sys.argv
-        new_environ = os.environ.copy()
-        new_environ["LARCH_BO_QT_FRONTEND_STARTED"] = "server"
-        server_process = Popen(args, env=new_environ, stderr=PIPE, close_fds=True, encoding="utf8")
-        url = server_process.stderr.readline().strip()
-        try:
+        if config.get("debug", False) and type_ != "debug":
+            from larch.bo.server.start import check_for_compile
+            config["root"] = root
+            check_for_compile(config)
+
+        url = "file://" + str(Path(resource_path)/"index.html")
+        if config.get("transmitter") or type_ == "debug":
+            port = get_free_port()
+            args = [sys.executable] + sys.argv + [f"--port={port}"]
+            new_environ = os.environ.copy()
+            new_environ["LARCH_BO_QT_FRONTEND_STARTED"] = "server"
+            new_environ["LARCH_START"] = "qt-server"
+            server_process = Popen(args, env=new_environ, encoding="utf8")
+            url += f"?transmitter={port}"
+            try:
+                return start_frontend(url, config, type_ == "debug")
+            finally:
+                server_process.terminate()
+        else:
             return start_frontend(url, config)
-        finally:
-            server_process.terminate()
